Moving some more methods away from daemon.go
